Restructuring and adding other pages
