+++ project/src/app/_page/sections/hero/index.tsx
