+++ project/src/app/_page/sections/mysticalPowers/FeatureCard.tsx
