+++ project/src/app/_page/sections/technologies/TechnologyCard.tsx
