+++ project/src/app/generate/page.tsx
@@ -1,0 +1,3 @@
+export default function Generate() {
+    return <></>;
+}
