+++ project/src/app/play/page.tsx
@@ -1,0 +1,3 @@
+export default function Play() {
+    return <></>;
+}
