+++ project/src/app/solve/page.tsx
@@ -1,0 +1,3 @@
+export default function Solve() {
+    return <></>;
+}
