+++ project/src/app/verify/page.tsx
@@ -1,0 +1,3 @@
+export default function Verify() {
+    return <></>;
+}
